test: moves tests into specified folders and chnages their namespaces
