+++ vCardLib.Tests/CollectionTests/EmailAddressCollectionTests.cs
@@ -3,7 +3,7 @@ using NUnit.Framework;
 using vCardLib.Collections;
 using vCardLib.Models;
 
-namespace vCardLib.Tests
+namespace vCardLib.Tests.CollectionTests
 {
 	[TestFixture]
 	public class EmailAddressCollectionTests
