+++ vCardLib.Tests/CollectionTests/ExpertiseCollectionTests.cs
@@ -1,10 +1,11 @@
 ﻿// Created by Bolorunduro Winner-Timothy on  11/18/2016 at 10:59 AM
+
 using System;
 using NUnit.Framework;
 using vCardLib.Collections;
 using vCardLib.Models;
 
-namespace vCardLib.Tests
+namespace vCardLib.Tests.CollectionTests
 {
 	[TestFixture]
 	public class ExpertiseCollectionTests
