+++ vCardLib.Tests/ModelTests/AddressTests.cs
@@ -1,7 +1,7 @@
 ﻿using NUnit.Framework;
 using vCardLib.Models;
 
-namespace vCardLib.Tests
+namespace vCardLib.Tests.ModelTests
 {
 	[TestFixture]
 	public class AddressTests
@@ -10,10 +10,12 @@ namespace vCardLib.Tests
 		public void AddressIsStable()
 		{
 			Assert.DoesNotThrow(delegate {
-				Address address = new Address();
-				address.Location = "Yaba Street";
-				address.Type = AddressType.Domestic;
-				address.Type = AddressType.Home;
+			    Address address = new Address
+			    {
+			        Location = "Yaba Street",
+			        Type = AddressType.Domestic
+			    };
+			    address.Type = AddressType.Home;
 				address.Type = AddressType.International;
 				address.Type = AddressType.None;
 				address.Type = AddressType.Parcel;
