+++ vCardLib.Tests/ModelTests/HobbyTests.cs
@@ -1,8 +1,9 @@
 ﻿// Created by Bolorunduro Winner-Timothy on  11/22/2016 at 5:18 AM
+
 using NUnit.Framework;
 using vCardLib.Models;
 
-namespace vCardLib.Tests
+namespace vCardLib.Tests.ModelTests
 {
 	[TestFixture]
 	public class HobbyTests
@@ -12,10 +13,12 @@ namespace vCardLib.Tests
 		{
 			Assert.DoesNotThrow(delegate
 			{
-				Hobby hobby = new Hobby();
-				hobby.Activity = "Tennis";
-				hobby.Level = Level.High;
-				hobby.Level = Level.Medium;
+			    Hobby hobby = new Hobby
+			    {
+			        Activity = "Tennis",
+			        Level = Level.High
+			    };
+			    hobby.Level = Level.Medium;
 				hobby.Level = Level.Low;
 			});
 		}
