+++ vCardLib.Tests/ModelTests/InterestTests.cs
@@ -1,8 +1,9 @@
 ﻿// Created by Bolorunduro Winner-Timothy on  11/22/2016 at 5:33 AM
+
 using NUnit.Framework;
 using vCardLib.Models;
 
-namespace vCardLib.Tests
+namespace vCardLib.Tests.ModelTests
 {
 	[TestFixture]
 	public class InterestTests
@@ -12,10 +13,12 @@ namespace vCardLib.Tests
 		{
 			Assert.DoesNotThrow(delegate
 			{
-				Interest interest = new Interest();
-				interest.Activity = "Tennis";
-				interest.Level = Level.High;
-				interest.Level = Level.Medium;
+			    Interest interest = new Interest
+			    {
+			        Activity = "Tennis",
+			        Level = Level.High
+			    };
+			    interest.Level = Level.Medium;
 				interest.Level = Level.Low;
 			});
 		}
